+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/annotation/VariantAnnotator.java
@@ -71,7 +71,7 @@ public final class VariantAnnotator {
 	/**
 	 * Convenience function for obtaining an {@link VariantAnnotations} from genome change in primitive types.
 	 *
-	 * Forwards to {@link #buildAnnotationList(int, int, String, String, PositionType)} and we recommend to use this
+	 * Forwards to {@link #buildAnnotations(int, int, String, String, PositionType)} and we recommend to use this
 	 * function directly.
 	 *
 	 * @param position
@@ -86,7 +86,7 @@ public final class VariantAnnotator {
 	 * @throws AnnotationException
 	 *             on problems building the annotation list
 	 */
-	public VariantAnnotations buildAnnotationList(int chr, int position, String ref, String alt, PositionType posType)
+	public VariantAnnotations buildAnnotations(int chr, int position, String ref, String alt, PositionType posType)
 			throws AnnotationException {
 		// Get chromosome by id.
 		if (chromosomeMap.get(chr) == null)
@@ -96,7 +96,7 @@ public final class VariantAnnotator {
 		GenomePosition pos = new GenomePosition(refDict, Strand.FWD, chr, position, posType);
 		GenomeVariant change = new GenomeVariant(pos, ref, alt);
 
-		return buildAnnotationList(change);
+		return buildAnnotations(change);
 	}
 
 	/**
@@ -111,7 +111,7 @@ public final class VariantAnnotator {
 	 * @throws AnnotationException
 	 *             on problems building the annotation list
 	 */
-	public VariantAnnotations buildAnnotationList(GenomeVariant change) throws AnnotationException {
+	public VariantAnnotations buildAnnotations(GenomeVariant change) throws AnnotationException {
 		// Short-circuit in the case of symbolic changes/alleles. These could be SVs, large duplications, etc., that are
 		// described as shortcuts in the VCF file. We cannot annotate these yet.
 		if (change.isSymbolic())